hal_core::traits

Trait PageTableTrait

pub trait PageTableTrait {
    // Required methods
    fn init(&mut self) -> Result<usize, ViError>;
    fn map(
        &mut self,
        virt: usize,
        phys: usize,
        flags: PageFlags,
        alloc_fn: &mut dyn FnMut() -> Option<usize>,
    ) -> Result<(), ViError>;
    fn unmap(&mut self, virt: usize) -> Result<(), ViError>;
    fn translate(&self, virt: usize) -> Option<usize>;
    unsafe fn activate(&self);

    // Provided method
    fn identity_map(
        &mut self,
        start: usize,
        end: usize,
        flags: PageFlags,
        alloc_fn: &mut dyn FnMut() -> Option<usize>,
    ) -> Result<(), ViError> { ... }
}

Generic Page Table Trait

Required Methods

fn init(&mut self) -> Result<usize, ViError>

Initialize paging (e.g., allocate root table). Typically returns the Physical Address of the root table.

fn map( &mut self, virt: usize, phys: usize, flags: PageFlags, alloc_fn: &mut dyn FnMut() -> Option<usize>, ) -> Result<(), ViError>

Map a virtual page to a physical frame. alloc_fn is a closure to allocate new frames for intermediate tables.

fn unmap(&mut self, virt: usize) -> Result<(), ViError>

Unmap a virtual page.

fn translate(&self, virt: usize) -> Option<usize>

Translate virtual address to physical address.

unsafe fn activate(&self)

Activate this page table (load into CR3/SATP).

Safety

Must ensure code/stack is mapped.

Provided Methods

fn identity_map( &mut self, start: usize, end: usize, flags: PageFlags, alloc_fn: &mut dyn FnMut() -> Option<usize>, ) -> Result<(), ViError>

Identity map a region (Virtual = Physical).